% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBalm.R
\name{createMappingHelpers}
\alias{createMappingHelpers}
\title{Helpers for parameterizing Lambda}
\usage{
createMappingHelpers(lform_output)
}
\value{
list
}
\description{
todo
}
\details{
mapping indicies to parameters
}

