\name{rBalm-package}
\alias{rBalm-package}
\alias{rBalm}
\docType{package}
\title{
\packageTitle{rBalm}
}
\description{
\packageDescription{rBalm}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{rBalm}
\packageIndices{rBalm}
}
\author{
\packageAuthor{rBalm}
Maintainer: \packageMaintainer{rBalm}}
\references{}
\keyword{package}
\seealso{}
\examples{}
