% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBalm.R
\name{rBalm}
\alias{rBalm}
\title{Robust Bayesian analysis of Luminex bead-based assays}
\usage{
rBalm(latent, tech = NULL, mfi_by, data, n_sample, n_thin, n_burn,
  update_beads = TRUE, prior_control = NULL)
}
\arguments{
\item{latent}{effect formula}

\item{mfi_by}{by statement for reducing data table}

\item{data}{data frame}

\item{n_thin}{thinning interval between samples}

\item{n_burn}{burn in samples}

\item{update_beads}{should bead-level parameters be updated?}

\item{prior_control}{list of named hyperparameters and values}

\item{technical}{effects random effect formula}

\item{n_iter}{number of samples to collect}
}
\value{
some output
}
\description{
todo description
}
\details{
todo details
}

