% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{rBalmInit}
\alias{rBalmInit}
\title{Initialize data structures for rBalm algorithm}
\usage{
rBalmInit(input, control)
}
\arguments{
\item{input}{A data.table with bead data}

\item{control}{A list with MCMC parameters}
}
\value{
List with required rBalm structures
}
\description{
todo
}
\details{
internal
}

