% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{processBeadVector}
\alias{processBeadVector}
\title{Process a bead distribution}
\usage{
processBeadVector(x)
}
\arguments{
\item{x}{bead values}
}
\value{
Return a named list with elements
}
\description{
Summary statistics and various quantities 
needed for Laplace density evaluation
}
\details{
The list is named
}

