\name{rBalm-package}
\alias{rBalm-package}
\alias{rBalm}
\docType{package}
\title{
\packageTitle{rBalm}
}
\description{
\packageDescription{rBalm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rBalm}
\packageIndices{rBalm}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rBalm}

Maintainer: \packageMaintainer{rBalm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
