% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBalm.R
\name{rBalm}
\alias{rBalm}
\title{Robust Bayesian analysis of Luminex bead-based assays}
\usage{
rBalm(data, n_iter, n_burn)
}
\arguments{
\item{data}{todo}

\item{n_iter}{todo}

\item{n_burn}{todo}
}
\value{
some output
}
\description{
todo description
}
\details{
todo details
}

